+++ deezy/audio_encoders/dee/base.py
@@ -1,19 +1,42 @@
 from abc import ABC, abstractmethod
 from enum import Enum
 from pathlib import Path
+import re
 import shutil
 import tempfile
 from typing import Generic, TypeVar
 
 from deezy.audio_encoders.base import BaseAudioEncoder
-from deezy.enums.shared import DeeFPS
+from deezy.audio_encoders.delay import get_dee_delay
+from deezy.enums.shared import DeeDelay, DeeFPS
 from deezy.exceptions import PathTooLongError
 from deezy.payloads.shared import ChannelBitrates
+from deezy.track_info.audio_track_info import AudioTrackInfo
+from deezy.utils.logger import logger
 
 DolbyChannelType = TypeVar("DolbyChannelType", bound=Enum)
 
 
 class BaseDeeAudioEncoder(BaseAudioEncoder, ABC, Generic[DolbyChannelType]):
+    def get_delay(
+        self,
+        audio_track_info: AudioTrackInfo,
+        payload_delay: str | None,
+        file_input: Path,
+    ) -> DeeDelay:
+        delay_str = "0ms"
+        # if audio is raw we'll parse delay from file if it exists
+        if audio_track_info.is_raw_audio:
+            delay_str = self._parse_delay_from_file(file_input)
+        # if delay is provided via payload (CLI) we'll override the above
+        if payload_delay:
+            delay_str = payload_delay
+        # generate dee delay
+        delay = get_dee_delay(delay_str)
+        if delay.is_delay():
+            logger.debug(f"Generated delay {delay.MODE}:{delay.DELAY}.")
+        return delay
+
     @staticmethod
     @abstractmethod
     def _get_channel_bitrate_object(
@@ -204,3 +227,11 @@ class BaseDeeAudioEncoder(BaseAudioEncoder, ABC, Generic[DolbyChannelType]):
         """
         if not keep_temp:
             shutil.rmtree(temp_dir)
+
+    @staticmethod
+    def _parse_delay_from_file(media_path: Path) -> str:
+        """Parse delay from filename, if None found return 0ms."""
+        match = re.search(r"delay\s*(-?\d+(?:ms|s))", media_path.name, flags=re.I)
+        if match:
+            return match.group(1)
+        return "0ms"